Fix 3

+++ 0x06-pointers_arrays_strings/3-strspn.c
@@ -7,7 +7,7 @@
 **/
 unsigned int _strspn(char *s, char *accept)
 {
-int i, ii;
+unsigned int i, ii;
 unsigned int o;
 
 for (i = 0; s[i] != '\0'; i++)
@@ -17,7 +17,7 @@ for (i = 0; s[i] != '\0'; i++)
 	if (s[i] == accept[ii])
 		o++;
 	}
-	if (i + 1 != (int) o)
+	if (i + 1 != o)
 		break;
 }
